create algolia config
