+++ src/eventshub/src/component/SearchTab/index.ts
@@ -1,0 +1,1 @@
+export * from './SearchTab'
